pyproject.toml
[project]
name = "git-report-generator"
requires-python = ">=3.11"
version = "0.1.0"
description = "A brief description of your project."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
dependencies = [
    "colorama>=0.4.6",
    "python-dotenv>=1.0.1",
]


